+++ scripts/admin.R
@@ -1,0 +1,7 @@
+
+# Packages ----------------------------------------------------------------
+
+renv::install("rappster/confx")
+usethis::use_package("confx")
+
+usethis::use_test("trans_add__generic")