+++ NEWS.md
@@ -1,3 +1,19 @@
+# trans 0.0.0.9002
+
+{confx}, config.yml and trans_add__generic()
+
+Fixed/closed:
+
+Changed:
+
+Added:
+
+- Package dependency `rappster/confx`
+- Config file `config.yml`
+- Function `trans_add__generic()` including unit test
+
+--------------------------------------------------------------------------------
+
 # trans 0.0.0.9001
 
 - Updated `.travis.yml`